backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.64.0
anyio==4.10.0
attrs==25.3.0
Authlib==1.6.1
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
coverage==7.10.3
cryptography==45.0.6
dataclasses-json==0.6.7
discord.py==2.5.2
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
fastapi==0.116.1
frozenlist==1.7.0
fsspec==2024.12.0
google-auth==2.40.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
json_repair==0.44.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-neo4j==0.5.0
langchain-openai==0.3.30
langchain-text-splitters==0.3.9
langsmith==0.4.14
MarkupSafe==3.0.2
marshmallow==3.26.1
motor==3.7.1
multidict==6.6.4
mypy_extensions==1.1.0
neo4j==5.28.2
neo4j-graphrag==1.9.1
numpy==2.3.2
openai==1.99.9
orjson==3.11.2
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.14.0
PyMuPDF==1.26.3
pypdf==5.9.0
pytesseract==0.3.13
pytest==8.4.1
pytest-cov==6.1.1
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rsa==4.9.1
scipy==1.16.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
starlette==0.47.2
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
types-PyYAML==6.0.12.20250809
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
yarl==1.20.1
zstandard==0.23.0
